firmware: added initial support for boot os cp/m from external usb storage
This is a work in progress feature.  If RomWBW or MSX ROM images are detected
at boot, expect unchanged behaviour.  Otherwise, system now attempt to load
and boot CP/M 2.2 OS from an external USB disk image.  Requires the Linear RAM
module for operation.

1. New top level directory - operating-system
2. New apps (via submodule xmodem80) XR.COM and XS.COM

See firmware/src/fbios/readme.txt for details of current progress

+++ firmware/src/rst-10-drivers/usb/base-drv/usb-base-drv.c
@@ -1,4 +1,3 @@
-#include "usb-base-drv.h"
 #include "../kyb-drv/kyb_driver.h"
 #include "../mse-drv/mse_driver.h"
 #include "ch376.h"
@@ -73,7 +72,7 @@ uint16_t usb_init(uint8_t state) {
   return (uint16_t)count_of_devices() << 8 | 4;
 }
 
-usb_error_t usb_scsi_seek(const uint16_t dev_index, const uint32_t lba) {
+usb_error_t usb_scsi_seek(const uint8_t dev_index, const uint32_t lba) {
   device_config_storage_t *const dev = (device_config_storage_t *)get_usb_device_config(dev_index);
 
   if (dev == NULL)
